+++ grails-app/domain/io/xh/hoist/jsonblob/JsonBlob.groovy
@@ -41,7 +41,17 @@ class JsonBlob implements JSONFormat {
         type maxSize: 50, blank: false
         owner maxSize: 50, nullable: true, blank: false
         acl nullable: true
-        name maxSize: 255, blank: false, unique: ['owner', 'type', 'archivedDate']
+        name maxSize: 255, blank: false, validator: { val, obj ->
+            if (where {
+                name == val &&
+                    type == obj.type &&
+                    archivedDate == obj.archivedDate &&
+                    owner == obj.owner &&
+                    token != obj.token
+            }.count() > 0) {
+                return 'default.not.unique.message'
+            }
+        }
         value validator: {Utils.isJSON(it) ?: 'default.invalid.json.message'}
         meta nullable: true, validator: {Utils.isJSON(it) ?: 'default.invalid.json.message'}
         description nullable: true